# swagger.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kasir Laundry Pro",
    "version": "1.0.0",
    "description": "API Routes and schema details of Kasir Laundry Pro Services"
  },
  "servers": [
    {
      "url": "https://kasirlaundrypro.netlify.app/api"
    }
  ],
  "paths": {
...
    "/customer": {
      "get": {
        "summary": "Get all customers",
        "description": "Retrieve a list of all customers",
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "Filter Data by name, phone number, email"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "description": "Create a new customer record",
        "tags": [
          "Customer"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input"
          }
        }
      }
    },
    "/customer/{id}": {
      "get": {
        "summary": "Get a customer by ID",
        "description": "Retrieve a customer record by ID",
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the customer to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "put": {
        "summary": "Update a customer",
        "description": "Update an existing customer record by ID",
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the customer to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a customer",
        "description": "Delete a customer record by ID",
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the customer to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Customer": {
        "type": "object",
        "required": [
          "id",
          "created_at",
          "name",
          "phone_number",
          "address",
          "gender"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the contact",
            "example": "2f19fb0e-2926-418d-82ac-c2dc33a0f4d3"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the contact was created",
            "example": "2024-08-07T13:58:23.934Z"
          },
          "name": {
            "type": "string",
            "description": "Name of the contact",
            "example": "sasaksj"
          },
          "phone_number": {
            "type": "string",
            "description": "Phone number of the contact",
            "example": "0238239238"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "description": "Email address of the contact (nullable)",
            "example": null
          },
          "address": {
            "type": "string",
            "description": "Address of the contact",
            "example": "kajsajs"
          },
          "gender": {
            "type": "string",
            "description": "Gender of the contact",
            "example": "Perempuan"
          }
        }
      },
      "CustomerRequestBody": {
        "type": "object",
        "required": [
          "name",
          "phone_number",
          "gender"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the customer",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "description": "Email address of the customer",
            "example": "john.doe@example.com"
          },
          "address": {
            "type": "string",
            "description": "Address of the customer",
            "example": "123 Main St, Springfield"
          },
          "phone_number": {
            "type": "string",
            "description": "Phone number of the customer, must be between 10 and 15 digits, and can optionally start with a plus sign",
            "pattern": "^[+]?[0-9]{10,15}$",
            "example": "+12345678901"
          },
          "gender": {
            "type": "string",
            "enum": [
              "Laki-laki",
              "Perempuan"
            ],
            "description": "Gender of the customer (Laki-laki or Perempuan)",
            "example": "Laki-laki"
          }
        }
      },
      "CustomerResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Customer created successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Customer"
          }
        }
      },
...
    }
  },
...
}